"Implemented navigation bar, ListItem component"
Added a navigation bar using the SearchBar component from the @rneui/themed library.
Created the ListItem component to represent individual items in our list.
Customized the background of the application.
Created a JSON file to store all the recipes, which will be used to display them in a list later.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: Tole15 <rodrigermoon@gmail.com>
Co-authored-by: BrandonVG <brandon.vadillox3@gmail.com>